requirements.txt
SQLAlchemy~=1.3.17
lxml~=4.5.0
python-dotenv~=0.13.0
requests~=2.23.0
selenium~=3.141.0
gevent~=20.5.1